+++ pygccxml/__init__.py
@@ -40,4 +40,4 @@ warnings.simplefilter("always", DeprecationWarning)
 # TODO:
 # 1. Add "explicit" property for constructors
 
-__version__ = '1.8.4'
+__version__ = '1.8.5'